pyproject.toml
[tool.poetry]
name = "students"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
setuptools = "^69.5.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
